Moved bluetooth to separate file, and clean up
